+++ main.tf
@@ -27,10 +27,11 @@ module "pipeline" {
   target_repo_arn    = module.codeartifact.repositories[var.pipeline.target_repository].arn
   domain_kms_key_arn = module.codeartifact.domain_kms_key_arn
 
-  cooldown = var.pipeline.cooldown
-  scanner  = var.pipeline.scanner
-  approval = var.pipeline.approval
-  logging  = var.pipeline.logging
+  cooldown       = var.pipeline.cooldown
+  scanner        = var.pipeline.scanner
+  approval       = var.pipeline.approval
+  logging        = var.pipeline.logging
+  yank_detection = var.pipeline.yank_detection
 
   log_retention_days = var.log_retention_days